requirements.txt
setuptools~=63.2.0
pandas~=1.4.3
selenium~=4.4.0
argparse~=1.4.0
numpy~=1.23.2
matplotlib~=3.6.2
Pillow~=9.3.0
nltk~=3.8
scikit-learn~=1.2.0
wordcloud~=1.8.2.2